+++ state.go
@@ -3,8 +3,6 @@ package zerotrace
 import (
 	"errors"
 	"fmt"
-	"math"
-	"math/rand"
 	"net"
 	"sync"
 	"time"
@@ -15,11 +13,6 @@ const (
 	ipidTimeout = time.Second * 10
 )
 
-var (
-	errInvalidResp = errors.New("got response for non-existing trace packet")
-	errNoMoreIds   = errors.New("all IP IDs are currently in flight")
-)
-
 // tracePkts represents a trace packet that we send to the client to determine
 // the network-level RTT.
 type tracePkt struct {
@@ -28,10 +21,6 @@ type tracePkt struct {
 	sent      time.Time
 	recvd     time.Time
 	recvdFrom net.IP
-}
-
-func (p *respPkt) String() string {
-	return fmt.Sprintf("<- %s (ttl=%d, ip id=%d)", p.recvdFrom.String(), p.ttl, p.ipID)
 }
 
 // respPkt represents a packet that we received in response to a trace packet.
@@ -57,9 +46,10 @@ func (p *tracePkt) String() string {
 //  2. Packets that we sent and received as part of the traceroute.
 //  3. The IP IDs that we use as part of the traceroute.
 type trState struct {
-	sync.RWMutex
-	dstAddr   net.IP
-	tracePkts map[uint16]*tracePkt
+	sync.Mutex // Guard tracePkts.
+	dstAddr    net.IP
+	tracePkts  map[uint16]*tracePkt
+	doneCh     chan struct{}
 }
 
 // newTrState returns a new traceroute state object.
@@ -67,63 +57,12 @@ func newTrState(dstAddr net.IP) *trState {
 	return &trState{
 		dstAddr:   dstAddr,
 		tracePkts: make(map[uint16]*tracePkt),
+		doneCh:    make(chan struct{}),
 	}
 }
 
-type ipIdState struct {
-	sync.Mutex // Guards ipids.
-	ipids      map[uint16]time.Time
-}
-
-func newIpIdState() *ipIdState {
-	return &ipIdState{
-		ipids: make(map[uint16]time.Time),
-	}
-}
-
-func (s *ipIdState) size() int {
-	s.Lock()
-	defer s.Unlock()
-
-	return len(s.ipids)
-}
-
-func (s *ipIdState) borrow() (uint16, error) {
-	s.Lock()
-	defer s.Unlock()
-
-	if len(s.ipids) == math.MaxUint16 {
-		return 0, errNoMoreIds
-	}
-
-	start := uint16(rand.Intn(math.MaxUint16))
-	for id := start + 1; id != start; id++ {
-		if _, exists := s.ipids[id]; !exists {
-			s.ipids[id] = time.Now().UTC()
-			return id, nil
-		}
-	}
-
-	return 0, nil
-}
-
-func (s *ipIdState) releaseUnanswered() {
-	s.Lock()
-	defer s.Unlock()
-
-	now := time.Now().UTC()
-	for id, added := range s.ipids {
-		if now.Sub(added) > ipidTimeout {
-			delete(s.ipids, id)
-		}
-	}
-}
-
-func (s *ipIdState) release(id uint16) {
-	s.Lock()
-	defer s.Unlock()
-
-	delete(s.ipids, id)
+func (s *trState) done() chan struct{} {
+	return s.doneCh
 }
 
 // AddTracePkt adds to the state map a trace packet.
@@ -136,27 +75,27 @@ func (s *trState) addTracePkt(p *tracePkt) {
 
 // AddRespPkt adds to the state map a packet that we got in response to a
 // previously-sent trace packet.
-func (s *trState) addRespPkt(p *respPkt) error {
+func (s *trState) addRespPkt(p *respPkt) {
 	s.Lock()
 	defer s.Unlock()
 
 	tracePkt, exists := s.tracePkts[p.ipID]
 	if !exists {
-		return errInvalidResp
+		return
 	}
 	// Mark the trace packet as "received".
 	tracePkt.recvd = p.recvd
 	tracePkt.recvdFrom = p.recvdFrom
-	return nil
+
+	if s.isFinished() {
+		close(s.doneCh)
+	}
 }
 
 // isFinished returns true if our state indicates that the 0trace scan is
 // finished.  That's the case when we haven't received any response packets
 // since the timeout.
 func (s *trState) isFinished() bool {
-	s.RLock()
-	defer s.RUnlock()
-
 	now := time.Now().UTC()
 	for _, p := range s.tracePkts {
 		if p.isAnswered() {
@@ -171,8 +110,8 @@ func (s *trState) isFinished() bool {
 
 // summary returns a printable string summary of the current traceroute state.
 func (s *trState) summary() string {
-	s.RLock()
-	defer s.RUnlock()
+	s.Lock()
+	defer s.Unlock()
 
 	numRcvd := 0
 	for _, p := range s.tracePkts {
@@ -189,8 +128,8 @@ func (s *trState) summary() string {
 // packet that made it the farthest to the client (i.e., the packet whose TTL
 // is the highest).
 func (s *trState) calcRTT() (time.Duration, error) {
-	s.RLock()
-	defer s.RUnlock()
+	s.Lock()
+	defer s.Unlock()
 
 	var closestPkt *tracePkt
 	for _, p := range s.tracePkts {
