+++ state_test.go
@@ -1,8 +1,6 @@
 package zerotrace
 
 import (
-	"errors"
-	"math"
 	"net"
 	"testing"
 	"time"
@@ -53,31 +51,6 @@ func TestAddTracePkt(t *testing.T) {
 	if len(s.tracePkts) != expected {
 		t.Fatalf("Expected %d recorded packets but got %d.",
 			expected, len(s.tracePkts))
-	}
-}
-
-func TestAddRespPkt(t *testing.T) {
-	s := newTrState(dummyAddr)
-	p := &respPkt{
-		ipID:      1,
-		ttl:       1,
-		recvd:     time.Now().UTC(),
-		recvdFrom: net.ParseIP("1.2.3.4"),
-	}
-
-	// Adding a packet in response to a non-existing trace packet should give
-	// us an error.
-	if err := s.addRespPkt(p); !errors.Is(err, errInvalidResp) {
-		t.Fatalf("Expected error %v but got %v.", errInvalidResp, err)
-	}
-
-	s.addTracePkt(&tracePkt{
-		ttl:  1,
-		ipID: 1,
-		sent: time.Now().UTC(),
-	})
-	if err := s.addRespPkt(p); err != nil {
-		t.Fatalf("Expected error nil but got %v.", err)
 	}
 }
 
@@ -201,50 +174,5 @@ func TestCalcRTT(t *testing.T) {
 	failOnErr(t, err)
 	if rtt != expectedRTT {
 		t.Fatalf("Expected RTT to be %s but got %s.", expectedRTT, rtt)
-	}
-}
-
-func TestExhaustGlobalState(t *testing.T) {
-	var (
-		err error
-		s   = newIpIdState()
-	)
-
-	// Exhaust the global state.
-	for i := 0; i < math.MaxUint16; i++ {
-		_, err = s.borrow()
-		failOnErr(t, err)
-	}
-	assertEqual(t, s.size(), math.MaxUint16)
-	// The global state is now full.  Subsequent requests for IP IDs should
-	// return an error.
-	_, err = s.borrow()
-	assertEqual(t, err, errNoMoreIds)
-}
-
-func TestGlobalState(t *testing.T) {
-	var (
-		numIDs = 100
-		ipids  = []uint16{}
-		s      = newIpIdState()
-	)
-
-	for i := 0; i < numIDs; i++ {
-		id, err := s.borrow()
-		failOnErr(t, err)
-		ipids = append(ipids, id)
-	}
-	assertEqual(t, s.size(), numIDs)
-
-	for _, id := range ipids {
-		s.release(id)
-	}
-	assertEqual(t, s.size(), 0)
-}
-
-func assertEqual(t *testing.T, is, should interface{}) {
-	t.Helper()
-	if should != is {
-		t.Fatalf("Expected value\n%v\nbut got\n%v", should, is)
 	}
 }
